+++ src/components/organisms/MobileSensorData/index.ts
@@ -1,1 +1,2 @@
 export { MobileSensorData } from "./MobileSensorData";
+export { useMobileSensorData } from "./MobileSensorData.hooks";
